requirements.txt
faiss-cpu
sentence_transformers
pandas
